Initial Progress Commit
First Commit. Currently this is the most promising of the 3 plants, and was the first plant to begin development when Rhynia began to take form. Will very likely be complete before the other 2. Still, due to significant changes, revisions, and improvements scheduled for it's parent mod Rhynia, will likely be a short while before it is ready.